requirement.txt
Django==3.2.4
django-cors-headers==3.7.0
bcrypt==3.2.0
PyJWT==2.3.0